+++ gar/include/aligator/gar/dense-riccati.hpp
@@ -22,7 +22,7 @@ public:
   using BlkMat44 = BlkMatrix<MatrixXs, 4, 4>;
   using BlkRowMat41 = BlkMatrix<RowMatrixXs, 4, 1>;
   using BlkVec4 = BlkMatrix<VectorXs, 4, 1>;
-  using KnotType = LQRKnotTpl<Scalar>;
+  using KnotType = LqrKnotTpl<Scalar>;
 
   std::vector<BlkMat44> kkts;
   std::vector<BlkVec4> ffs;