+++ gar/include/aligator/gar/fwd.hpp
@@ -6,7 +6,7 @@ namespace aligator {
 namespace gar {
 // FORWARD DECLARATIONS
 
-template <typename Scalar> struct LQRKnotTpl;
+template <typename Scalar> struct LqrKnotTpl;
 template <typename Scalar> struct LQRProblemTpl;
 template <typename Scalar> class RiccatiSolverBase;
 } // namespace gar