+++ gar/include/aligator/gar/utils.hxx
@@ -10,7 +10,7 @@ void lqrCreateSparseMatrix(const LQRProblemTpl<Scalar> &problem,
                            Eigen::Matrix<Scalar, -1, 1> &rhs, bool update) {
   using Eigen::Index;
   const uint nrows = lqrNumRows(problem);
-  using knot_t = LQRKnotTpl<Scalar>;
+  using knot_t = LqrKnotTpl<Scalar>;
   const auto &knots = problem.stages;
 
   if (!update) {
@@ -106,7 +106,7 @@ std::array<Scalar, 3> lqrComputeKktError(
   uint N = (uint)problem.horizon();
   assert(xs.size() == N + 1);
   using VectorXs = typename math_types<Scalar>::VectorXs;
-  using KnotType = LQRKnotTpl<Scalar>;
+  using KnotType = LqrKnotTpl<Scalar>;
 
   Scalar dynErr = 0.;
   Scalar cstErr = 0.;
@@ -195,7 +195,7 @@ template <typename Scalar>
 bool lqrDenseMatrix(const LQRProblemTpl<Scalar> &problem, Scalar mudyn,
                     Scalar mueq, typename math_types<Scalar>::MatrixXs &mat,
                     typename math_types<Scalar>::VectorXs &rhs) {
-  using knot_t = LQRKnotTpl<Scalar>;
+  using knot_t = LqrKnotTpl<Scalar>;
   const auto &knots = problem.stages;
   const size_t N = size_t(problem.horizon());